+++ modules/terraform-aci-access-leaf-interface-profile/examples/complete/main.tf
@@ -2,5 +2,6 @@ module "aci_access_leaf_interface_profile" {
   source  = "netascode/nac-aci/aci//modules/terraform-aci-access-leaf-interface-profile"
   version = ">= 0.8.0"
 
-  name = "INT-PROF1"
+  name        = "INT-PROF1"
+  description = "Interface profile 101"
 }